% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inegi.R
\name{choropleth_inegi}
\alias{choropleth_inegi}
\title{choropleth_inegi}
\usage{
choropleth_inegi(token, regions, indicator, title, legend = "",
  silent = FALSE)
}
\arguments{
\item{token}{The INEGI registration token available from http://www3.inegi.org.mx/sistemas/api/denue/v1/tokenVerify.aspx}

\item{regions}{A vector of states to zoom in on. Elements of this vector must exactly match the names of states as they appear in the "region" column of ?df_mxstates.}

\item{indicator}{Numeric indicator}

\item{title}{An optional title for the map.}

\item{legend}{An optional name for the legend.}

\item{silent}{Print a text progress bar with estimated time remaining}
}
\value{
A choropleth
}
\description{
choropleth_inegi
}
\examples{
\dontrun{
data(df_mxstates)
state_regions <- df_mxstate$region
choropleth_inegi(token, state_regions, "3101008001")
}
}

