% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxmaps-package.R
\name{choroplethr_animate}
\alias{choroplethr_animate}
\title{Animate a list of chropleths}
\arguments{
\item{choropleths}{A list of choropleths represented as ggplot2 objects.}
}
\value{
Nothing.  However, a variable number of files are written to the current working directory.
}
\description{
Animate a list of chropleths
}
