% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_mxmun}
\alias{df_mxmun}
\title{A data.frame containing population estimates for Mexican municipios in 2015.}
\usage{
data(df_mxmun)
}
\description{
A data.frame containing population estimates for Mexican municipios in 2015.
}
\references{
Taken from the Encuesta Intercensal..
}
\keyword{data}

