% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_mxstate}
\alias{df_mxstate}
\title{Mexican states dataset}
\format{
An object of class \code{data.frame} with 32 rows and 13 columns.
}
\usage{
data(df_mxstate)
}
\description{
A data.frame containing population estimates for all the Mexican states in 2015.
}
\details{
\describe{
  \item{region}{INEGI code of the state}
  \item{state_name}{short state name (e.g. Coahuila)}
  \item{state_name_official}{Official state name (e.g. Coahuila de Zaragoza)}
  \item{state_abbr}{state abbreviation}
  \item{state_abbr_official}{official state abbreviation (it can be awkward to use Chis for Chiapas)
  according to the INEGI.}
  \item{year}{2015, the year of the Conteo from which the data is sourced}
  \item{pop}{total state population according to the Encuesta Intercensal 2015}
  \item{pop_male}{male population according to the Encuesta Intercensal 2015}
  \item{pop_female}{female population according to the Encuesta Intercensal 2015}
  \item{afromexican}{afromexican population according to the Encuesta Intercensal 2015}
  \item{part_afromexican}{part afromexican population according to the Encuesta Intercensal 2015}
  \item{indigenous}{indigenous population according to the Encuesta Intercensal 2015}
  \item{part_indigenous}{part indigenous population according to the Encuesta Intercensal 2015}
}
}
\examples{
data("df_mxstate")
head(df_mxstate)
}
\references{
Population estimates taken from the
\href{https://www.inegi.org.mx/programas/intercensal/2015/}{Encuesta Intercensal.}
}
\keyword{data}
