% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inegi.R
\name{get_regions_inegi}
\alias{get_regions_inegi}
\title{Title}
\usage{
get_regions_inegi(token, regions, indicator, silent = TRUE)
}
\arguments{
\item{token}{The INEGI registration token available from http://www3.inegi.org.mx/sistemas/api/denue/v1/tokenVerify.aspx}

\item{regions}{Region to download data}

\item{indicator}{Numeric indicator}

\item{silent}{print progress}
}
\description{
Title
}
