% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inegi.R
\name{hexbin_inegi}
\alias{hexbin_inegi}
\title{Hexbin map with INEGI data}
\usage{
hexbin_inegi(token, regions, indicator, title, legend = "",
  database = "BISE", silent = FALSE)
}
\arguments{
\item{token}{The INEGI registration token available from http://www3.inegi.org.mx/sistemas/api/denue/v1/tokenVerify.aspx}

\item{regions}{A vector of states to zoom in on. Elements of this vector must exactly match the names of states as they appear in the "region" column of ?df_mxstates.}

\item{indicator}{Numeric indicator from the \href{http://www.beta.inegi.org.mx/servicios/api_biinegi.html}{INEGI API}}

\item{title}{An optional title for the map.}

\item{legend}{An optional name for the legend.}

\item{database}{The database to query. BIE (Banco de Informacion Economica) or BISE (Banco de Indicadores). Defaults to BIE.}

\item{silent}{Wether to print a text progress bar while downloading the INEGI API data}
}
\value{
A choropleth
}
\description{
A state hexbin map of data available from the INEGI API
}
\examples{
\dontrun{
data(df_mxstates)
## Insert token here
## If you don't have a token you can get one from:
## http://www3.inegi.org.mx//sistemas/api/indicadores/v1/tokenVerify.aspx
token <- "INSERT_TOKEN_HERE"
state_regions <- df_mxstate$region
## You can look up the numeric indicator codes at
## http://www.beta.inegi.org.mx/servicios/api_biinegi.html
hexbin_inegi(token, state_regions, "6200205259")
}
}
