% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\name{mxhexbin_choropleth}
\alias{mxhexbin_choropleth}
\title{Create a state-level hexbin choropleth}
\usage{
mxhexbin_choropleth(
  df,
  title = "",
  legend = "",
  state_border_color = "dark grey",
  state_border_size = 0.2,
  background_color = "white",
  title_color = "black",
  title_align = 0.5,
  title_position = "plot",
  num_colors = 7,
  zoom = NULL,
  label_color = "black",
  label_size = 4.5,
  shadow_color = NULL,
  auto_contrast = NULL
)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".  Elements in
the "region" column must exactly match how regions are named in the "region" column in ?df_mxstate.}

\item{title}{An optional title for the map.}

\item{legend}{An optional name for the legend.}

\item{state_border_color}{Border color of polygons.}

\item{state_border_size}{Border line width.}

\item{background_color}{Background color of the map and legend.}

\item{title_color}{Title text color.}

\item{title_align}{Horizontal justification of title.}

\item{title_position}{Either "plot" or "panel".}

\item{num_colors}{The number of colors to use on the map.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many colors.}

\item{zoom}{An optional vector of states to zoom in on. Elements of this vector must exactly
match the names of countries as they appear in the "region" column of ?country.regions}

\item{label_color}{An optional color for the state abbreviation labels}

\item{label_size}{An optional size for the state abbrevition labels}

\item{shadow_color}{An optional shadow color for the state abbreviation labels}

\item{auto_contrast}{Optional argument to automatically determine the best contrast
for the state abbreviation labels}
}
\value{
A ggplot object representing an hexbin of Mexico
}
\description{
Hexagonal tiles of the states of Mexico
}
\examples{
data(df_mxstate)
df_mxstate$value <- df_mxstate$pop
mxhexbin_choropleth(df_mxstate, num_colors = 1)
}
