% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxmaps-package.R
\docType{package}
\name{mxmaps}
\alias{mxmaps-package}
\alias{mxmaps}
\title{\code{mxmaps} package}
\description{
Tools for static and interactive choropleths of Mexico. Includes functions to
manipulate INEGI state and municipio codes along with options to download data
from the INEGI API.
}
\details{
See the website on
\href{https://www.diegovalle.net/mxmaps/}{mxmaps}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.diegovalle.net/mxmaps/}
  \item \url{https://github.com/diegovalle/mxmaps}
  \item Report bugs at \url{https://github.com/diegovalle/mxmaps/issues}
}

}
\author{
\strong{Maintainer}: Diego Valle-Jones \email{diego@diegovalle.net}

Authors:
\itemize{
  \item Diego Valle-Jones \email{diego@diegovalle.net}
}

}
