% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{str_mxmunicipio}
\alias{str_mxmunicipio}
\title{Format municipio codes with leading zeroes}
\usage{
str_mxmunicipio(code, municipio_code)
}
\arguments{
\item{code}{The sate or combined state and municipio code if the municipio_code variable is missing}

\item{municipio_code}{The municipio code}
}
\value{
The state and municipio code in the format required by the plotting functions
}
\description{
Format municipio codes with leading zeroes
}
\examples{
# Format the concatenated string of state code + municipio code
str_mxmunicipio(c("1006", "2003"))
# Format the state and municipio code separately
str_mxmunicipio(c(10, 6), c(20, 3))

\dontrun{
# warning about invalid code
str_mxmunicipio(33, 999)
}
}
\seealso{
\url{http://www.inegi.org.mx/est/contenidos/proyectos/aspectosmetodologicos/clasificadoresycatalogos/catalogo_entidades.aspx}
}
