% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{str_mxstate}
\alias{str_mxstate}
\title{str_mxstate}
\usage{
str_mxstate(code)
}
\arguments{
\item{code}{The state code according to the INEGI's Catalogo de entidades federativas, municipios y localidades}
}
\value{
The state code in the format required by the plotting functions
}
\description{
str_mxstate
}
\examples{
str_mxstate(c("12", "2"))
str_mxstate(c(32, 5))

\dontrun{
# warning about invalid state code
str_mxstate("35")
}
}
\seealso{
\url{http://www.inegi.org.mx/est/contenidos/proyectos/aspectosmetodologicos/clasificadoresycatalogos/catalogo_entidades.aspx}
}
