% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{Choropleth2}
\alias{Choropleth2}
\title{The base Choropleth2 object.}
\description{
The base Choropleth2 object.

The base Choropleth2 object.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{user.df}}{input from user}

\item{\code{map.df}}{geometry of the map}

\item{\code{choropleth.df}}{result of binding user data with our map data}

\item{\code{title}}{title for map}

\item{\code{legend}}{title for legend}

\item{\code{warn}}{warn user on clipped or missing values}

\item{\code{ggplot_scale}}{override default scale.}

\item{\code{ggplot_polygon}}{null}

\item{\code{projection_sf}}{null}

\item{\code{projection}}{null}

\item{\code{ggplot_sf}}{null}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Choropleth2-new}{\code{Choropleth2$new()}}
\item \href{#method-Choropleth2-render}{\code{Choropleth2$render()}}
\item \href{#method-Choropleth2-get_min_long}{\code{Choropleth2$get_min_long()}}
\item \href{#method-Choropleth2-get_max_long}{\code{Choropleth2$get_max_long()}}
\item \href{#method-Choropleth2-get_min_lat}{\code{Choropleth2$get_min_lat()}}
\item \href{#method-Choropleth2-get_max_lat}{\code{Choropleth2$get_max_lat()}}
\item \href{#method-Choropleth2-get_bounding_box}{\code{Choropleth2$get_bounding_box()}}
\item \href{#method-Choropleth2-get_x_scale}{\code{Choropleth2$get_x_scale()}}
\item \href{#method-Choropleth2-get_y_scale}{\code{Choropleth2$get_y_scale()}}
\item \href{#method-Choropleth2-get_reference_map}{\code{Choropleth2$get_reference_map()}}
\item \href{#method-Choropleth2-get_choropleth_as_polygon}{\code{Choropleth2$get_choropleth_as_polygon()}}
\item \href{#method-Choropleth2-render_with_reference_map}{\code{Choropleth2$render_with_reference_map()}}
\item \href{#method-Choropleth2-clip}{\code{Choropleth2$clip()}}
\item \href{#method-Choropleth2-discretize}{\code{Choropleth2$discretize()}}
\item \href{#method-Choropleth2-bind}{\code{Choropleth2$bind()}}
\item \href{#method-Choropleth2-prepare_map}{\code{Choropleth2$prepare_map()}}
\item \href{#method-Choropleth2-get_scale}{\code{Choropleth2$get_scale()}}
\item \href{#method-Choropleth2-theme_clean}{\code{Choropleth2$theme_clean()}}
\item \href{#method-Choropleth2-theme_inset}{\code{Choropleth2$theme_inset()}}
\item \href{#method-Choropleth2-format_levels}{\code{Choropleth2$format_levels()}}
\item \href{#method-Choropleth2-set_zoom}{\code{Choropleth2$set_zoom()}}
\item \href{#method-Choropleth2-get_zoom}{\code{Choropleth2$get_zoom()}}
\item \href{#method-Choropleth2-set_num_colors}{\code{Choropleth2$set_num_colors()}}
\item \href{#method-Choropleth2-clone}{\code{Choropleth2$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-new"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-new}{}}}
\subsection{Method \code{new()}}{
a choropleth map is defined by these two variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$new(map.df, user.df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{map.df}}{a data.frame of a map}

\item{\code{user.df}}{a data.frame that expresses values for regions of each map}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-render"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-render}{}}}
\subsection{Method \code{render()}}{
render
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$render()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-get_min_long"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-get_min_long}{}}}
\subsection{Method \code{get_min_long()}}{
left
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$get_min_long()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-get_max_long"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-get_max_long}{}}}
\subsection{Method \code{get_max_long()}}{
right
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$get_max_long()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-get_min_lat"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-get_min_lat}{}}}
\subsection{Method \code{get_min_lat()}}{
bottom
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$get_min_lat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-get_max_lat"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-get_max_lat}{}}}
\subsection{Method \code{get_max_lat()}}{
top
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$get_max_lat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-get_bounding_box"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-get_bounding_box}{}}}
\subsection{Method \code{get_bounding_box()}}{
get_bounding_box
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$get_bounding_box(long_margin_percent, lat_margin_percent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{long_margin_percent}}{null}

\item{\code{lat_margin_percent}}{null}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-get_x_scale"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-get_x_scale}{}}}
\subsection{Method \code{get_x_scale()}}{
get_x_scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$get_x_scale()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-get_y_scale"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-get_y_scale}{}}}
\subsection{Method \code{get_y_scale()}}{
get_y_scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$get_y_scale()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-get_reference_map"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-get_reference_map}{}}}
\subsection{Method \code{get_reference_map()}}{
get_reference_map
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$get_reference_map()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-get_choropleth_as_polygon"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-get_choropleth_as_polygon}{}}}
\subsection{Method \code{get_choropleth_as_polygon()}}{
get_choropleth_as_polygon
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$get_choropleth_as_polygon(alpha)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{null}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-render_with_reference_map"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-render_with_reference_map}{}}}
\subsection{Method \code{render_with_reference_map()}}{
render_with_reference_map
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$render_with_reference_map(alpha = 0.5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{null}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-clip"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-clip}{}}}
\subsection{Method \code{clip()}}{
support e.g. users just viewing states on the west coast
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$clip()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-discretize"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-discretize}{}}}
\subsection{Method \code{discretize()}}{
discretizing values means
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$discretize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-bind"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-bind}{}}}
\subsection{Method \code{bind()}}{
bind
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$bind()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-prepare_map"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-prepare_map}{}}}
\subsection{Method \code{prepare_map()}}{
prepare_map
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$prepare_map()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-get_scale"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-get_scale}{}}}
\subsection{Method \code{get_scale()}}{
get_scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$get_scale()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-theme_clean"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-theme_clean}{}}}
\subsection{Method \code{theme_clean()}}{
theme_clean
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$theme_clean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-theme_inset"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-theme_inset}{}}}
\subsection{Method \code{theme_inset()}}{
theme_inset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$theme_inset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-format_levels"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-format_levels}{}}}
\subsection{Method \code{format_levels()}}{
format_levels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$format_levels(x, nsep = " to ")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{null}

\item{\code{nsep}}{null}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-set_zoom"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-set_zoom}{}}}
\subsection{Method \code{set_zoom()}}{
set_zoom
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$set_zoom(zoom)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{zoom}}{null}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-get_zoom"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-get_zoom}{}}}
\subsection{Method \code{get_zoom()}}{
get_zoom
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$get_zoom()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-set_num_colors"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-set_num_colors}{}}}
\subsection{Method \code{set_num_colors()}}{
set_num_colors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$set_num_colors(num_colors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_colors}}{null}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Choropleth2-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Choropleth2-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Choropleth2$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
