% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_1990_2010.R
\docType{data}
\name{df_mxmunicipio_1990_2010}
\alias{df_mxmunicipio_1990_2010}
\title{Mexican 1990-2010 municipio population dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36855 rows and 5 columns.
}
\usage{
data(df_mxmunicipio_1990_2010)
}
\description{
A data.frame containing population estimates for all the Mexican municipios in
1990, 1995, 2000, 2005, and 2010
}
\details{
\describe{
  \item{region}{INEGI code of the state}
  \item{municipio_name}{official name of each municipio}
  \item{year}{2015, the year of the Conteo from which the data is sourced}
  \item{sex}{Sex of the person}
  \item{pop}{total municipio population}
}
}
\examples{
data("df_mxmunicipio_1990_2010")
head(df_mxmunicipio_1990_2010)
}
\references{
Population estimates taken from the \href{https://www.inegi.org.mx/programas/ccpv/2020/default.html#Tabulados}{INEGI Census Page}
}
\keyword{data}
