% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2020.R
\docType{data}
\name{df_mxmunicipio_2020}
\alias{df_mxmunicipio_2020}
\title{Mexican 2020 municipio dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2469 rows and 17 columns.
}
\usage{
data(df_mxmunicipio_2020)
}
\description{
A data.frame containing population estimates for all the Mexican municipios in 2020
}
\details{
\describe{
  \item{state_code}{INEGI code of each state}
  \item{municipio_code}{INEGI code of each municipio}
  \item{region}{INEGI code of the state}
  \item{state_name}{short state name (e.g. Coahuila)}
  \item{state_name_official}{Official state name (e.g. Coahuila de Zaragoza)}
  \item{state_abbr}{state abbreviation}
  \item{state_abbr_official}{official state abbreviation (it can be awkward to use Chis for Chiapas)
  according to the INEGI.}
  \item{municipio_name}{official name of each municipio}
  \item{year}{2015, the year of the Conteo from which the data is sourced}
  \item{pop}{total municipio population according to the Censo 2020}
  \item{pop_male}{male population according to the Censo 2020}
  \item{pop_female}{female population according to the Censo 2020}
  \item{afromexican}{afromexican population according to the Censo 2020}
  \item{indigenous_language}{Number of persons who speak an indigenous language
  according to the Censo 2020}
  \item{long}{longitude of the localidad cabecera of each municipio}
  \item{lat}{latitude of the localidad cabecera of each municipio}
}
}
\examples{
data("df_mxmunicipio_2020")
head(df_mxmunicipio_2020)
}
\references{
Population estimates taken from the
\href{https://www.inegi.org.mx/programas/ccpv/2020/default.html#Tabulados}{Censo 2020}.
The latitude and longitude of the localidad cabecera of each municipio come from the
\href{https://www.inegi.org.mx/app/ageeml/}{Catálogo Único de Claves
de Áreas Geoestadísticas Estatales, Municipales y Localidades}
}
\keyword{data}
