% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_mxstate}
\alias{df_mxstate}
\title{A data.frame containing population estimates for Mexican states in 2015.}
\usage{
data(df_mxstate)
}
\description{
A data.frame containing population estimates for Mexican states in 2015.
}
\references{
Taken from the Encuesta Intercensal.
}
\keyword{data}
