% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_1990_2010.R
\docType{data}
\name{df_mxstate_1990_2010}
\alias{df_mxstate_1990_2010}
\title{Mexican 1990-2010 states population dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 480 rows and 5 columns.
}
\usage{
df_mxstate_1990_2010
}
\description{
A data.frame containing population estimates for all the Mexican states in
1990, 1995, 2000, 2005, and 2010
}
\details{
\describe{
  \item{region}{INEGI code of the state}
  \item{state_name_official}{Official state name (e.g. Coahuila de Zaragoza)}
  \item{year}{2015, the year of the Conteo from which the data is sourced}
  \item{sex}{Sex of the person}
  \item{pop}{total state population}
}
}
\examples{
data("df_mxstate_1990_2010")
head(df_mxstate_1990_2010)
}
\references{
Population estimates taken from the \href{https://www.inegi.org.mx/programas/ccpv/2020/default.html#Tabulados}{INEGI Census Page.}
}
\keyword{data}
