% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inegi.R
\name{.get_inegi_data}
\alias{.get_inegi_data}
\title{Title}
\usage{
.get_inegi_data(token, region, indicator, database)
}
\arguments{
\item{token}{The INEGI registration token available from http://www3.inegi.org.mx/sistemas/api/denue/v1/tokenVerify.aspx}

\item{region}{Region to download data}

\item{indicator}{Numeric indicator}

\item{database}{The database to query. BIE (Banco de Informacion Economica) or BISE (Banco de Indicadores). Defaults to BIE.}
}
\description{
Title
}
\keyword{internal}
