% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin_data.R
\docType{data}
\name{mxhexbin.map}
\alias{mxhexbin.map}
\title{Hexbin map of Mexican states}
\usage{
data(mxhexbin.map)
}
\description{
A fortified data.frame which contains an hexagonal bin map of all 32 Mexican states
}
\examples{
## render the map with ggplot2
library(ggplot2)

data(mxhexbin.map)
ggplot(mxhexbin.map, aes(long, lat, group=group)) +
  geom_polygon(color = "black") +
  coord_map()
}
