% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin_data.R
\docType{data}
\name{mxhexbin.map}
\alias{mxhexbin.map}
\title{Hex bin map of the 31 Mexican states plus the Federal District}
\usage{
data(mxhexbin.map)
}
\description{
A data.frame which contains an hexagonal bin map of all 31 Mexican States plus
the Federal District.
}
\examples{
\dontrun{
# render the map with ggplot2
library(ggplot2)

data(mxhexbin.map)
ggplot(mxhexbin.map, aes(long, lat, group=group)) + geom_polygon(color = "black")
}
}
