% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\name{mxhexbin_choropleth}
\alias{mxhexbin_choropleth}
\title{Create a municipio-level choropleth}
\usage{
mxhexbin_choropleth(df, title = "", legend = "", num_colors = 7,
  zoom = NULL, label_color = "black", label_size = 4.5)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".  Elements in
the "region" column must exactly match how regions are named in the "region" column in ?country.map.}

\item{title}{An optional title for the map.}

\item{legend}{An optional name for the legend.}

\item{num_colors}{The number of colors to use on the map.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many colors.}

\item{zoom}{An optional vector of countries to zoom in on. Elements of this vector must exactly
match the names of countries as they appear in the "region" column of ?country.regions}

\item{label_color}{An optional color for the state abbreviation labels}

\item{label_size}{An optional size for the state abbrevition labels}
}
\description{
The map used is mxmunicipio.map in the mxmapsData package. See mxmunicipio.map for
an object which can help you coerce your regions into the required format.
}
\examples{
\dontrun{
data(df_mxstates)
df_mxstates$value <- df_mxstates$pop
mxhexbin _choropleth(df_mxstates, num_colors = 1)
}
}
