% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\name{MXHexBinChoropleth}
\alias{MXHexBinChoropleth}
\title{An R6 object for creating state-level hexbin choropleths.}
\description{
An R6 object for creating state-level hexbin choropleths.
}
\examples{
library(viridis)
library(scales)

df_mxstate$value <-  df_mxstate$indigenous / df_mxstate$pop
gg = MXHexBinChoropleth$new(df_mxstate)
gg$title <- "Percentage of the population that self-identifies as indigenous"
gg$set_num_colors(1)
gg$ggplot_scale <- scale_fill_viridis("percent", labels = percent)
gg$render()
}
\section{Super class}{
\code{\link[mxmaps:Choropleth2]{Choropleth2}} -> \code{MXHexBinChoropleth}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{label_color}}{color for the state labels}

    \item{\code{shadow_color}}{shadow color for the state labels}

    \item{\code{label_size}}{font size for the state labels}

    \item{\code{show_labels}}{draw the state labels}

    \item{\code{auto_contrast}}{automatically determine the label color based on the map fill scale}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-MXHexBinChoropleth-initialize}{\code{MXHexBinChoropleth$new()}}
    \item \href{#method-MXHexBinChoropleth-render}{\code{MXHexBinChoropleth$render()}}
    \item \href{#method-MXHexBinChoropleth-clone}{\code{MXHexBinChoropleth$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="bind"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-bind'><code>Choropleth2$bind()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="clip"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-clip'><code>Choropleth2$clip()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="discretize"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-discretize'><code>Choropleth2$discretize()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="format_levels"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-format_levels'><code>Choropleth2$format_levels()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="get_bounding_box"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-get_bounding_box'><code>Choropleth2$get_bounding_box()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="get_choropleth_as_polygon"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-get_choropleth_as_polygon'><code>Choropleth2$get_choropleth_as_polygon()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="get_max_lat"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-get_max_lat'><code>Choropleth2$get_max_lat()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="get_max_long"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-get_max_long'><code>Choropleth2$get_max_long()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="get_min_lat"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-get_min_lat'><code>Choropleth2$get_min_lat()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="get_min_long"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-get_min_long'><code>Choropleth2$get_min_long()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="get_reference_map"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-get_reference_map'><code>Choropleth2$get_reference_map()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="get_scale"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-get_scale'><code>Choropleth2$get_scale()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="get_x_scale"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-get_x_scale'><code>Choropleth2$get_x_scale()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="get_y_scale"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-get_y_scale'><code>Choropleth2$get_y_scale()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="get_zoom"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-get_zoom'><code>Choropleth2$get_zoom()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="prepare_map"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-prepare_map'><code>Choropleth2$prepare_map()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="render_with_reference_map"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-render_with_reference_map'><code>Choropleth2$render_with_reference_map()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="set_num_colors"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-set_num_colors'><code>Choropleth2$set_num_colors()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="set_zoom"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-set_zoom'><code>Choropleth2$set_zoom()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="theme_clean"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-theme_clean'><code>Choropleth2$theme_clean()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mxmaps" data-topic="Choropleth2" data-id="theme_inset"><a href='../../mxmaps/html/Choropleth2.html#method-Choropleth2-theme_inset'><code>Choropleth2$theme_inset()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MXHexBinChoropleth-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-MXHexBinChoropleth-initialize}{}}}
\subsection{\code{MXHexBinChoropleth$new()}}{
  Initialize the map of Mexico
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MXHexBinChoropleth$new(user.df)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{user.df}}{df}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new `MXHexBinChoropleth` object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MXHexBinChoropleth-render"></a>}}
\if{latex}{\out{\hypertarget{method-MXHexBinChoropleth-render}{}}}
\subsection{\code{MXHexBinChoropleth$render()}}{
  Render a map of Mexico
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MXHexBinChoropleth$render()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A ggplot2 object with the map of Mexico.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MXHexBinChoropleth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MXHexBinChoropleth-clone}{}}}
\subsection{\code{MXHexBinChoropleth$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MXHexBinChoropleth$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
