% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\docType{data}
\name{MXHexBinChoropleth}
\alias{MXHexBinChoropleth}
\title{An R6 object for creating municipio-level choropleths.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MXHexBinChoropleth
}
\description{
An R6 object for creating municipio-level choropleths.
}
\keyword{datasets}
