% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxmaps.R
\docType{package}
\name{mxmaps}
\alias{mxmaps}
\alias{mxmaps-package}
\title{\code{mxmaps} package}
\description{
Tools for static and interactive choropleths of Mexico. Includes functions to
manipulate INEGI state and municipio codes along with options to download data
from the INEGI API.
}
\details{
See the website on
\href{https://www.diegovalle.net/mxmaps/}{mxmaps}
}
