% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipio_data.R
\docType{data}
\name{mxmunicipio.map}
\alias{mxmunicipio.map}
\title{Map of the all Mexican municipios and alcaldías}
\usage{
data(mxmunicipio.map)
}
\description{
A data.frame which contains a map of all Mexican municipios plus
boroughs of the Federal District.  The shapefile
was modified using QGIS in order to remove
the Maria islands, Socorro island, Arrecife Alacran and
Guadalupe Island, then simplified with http://www.mapshaper.org/
}
\examples{
# render the map with ggplot2
library(ggplot2)

data(mxmunicipio.map)
ggplot(mxmunicipio.map, aes(long, lat, group=group)) +
  geom_polygon(fill = "white", color = "#666666", size = .05) +
  coord_map()
}
\references{
Downloaded from the "Cartografia Geoestadistica Urbana y
Rural Amanzanada. Planeacion de la Encuesta Intercensal 2015" shapefiles
(https://gist.github.com/diegovalle/aa3eef87c085d6ea034f)
}
