% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoJSON_data.R
\docType{data}
\name{mxmunicipio.topoJSON}
\alias{mxmunicipio.topoJSON}
\title{Map of the all Mexican municipios and delegaciones}
\usage{
data(mxmunicipio.topoJSON)
}
\description{
A data.frame which contains a map of all Mexican municipios plus
boroughs of the Federal District in topoJSON format.
}
\references{
Downloaded from the "Cartografia Geoestadistica Urbana y
Rural Amanzanada. Planeacion de la Encuesta Intercensal 2015" shapefiles
(https://gist.github.com/diegovalle/aa3eef87c085d6ea034f)
}

