% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipio.R
\name{mxmunicipio_choropleth}
\alias{mxmunicipio_choropleth}
\title{Create a municipio-level choropleth}
\usage{
mxmunicipio_choropleth(
  df,
  title = "",
  legend = "",
  num_colors = 7,
  zoom = NULL,
  show_states = TRUE
)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".  Elements in
the "region" column must match the numeric codes in the "region" column of ?df_mxmunicipio
either with a leading zero or without one (e.g. 01001 or 1001 are both fine)}

\item{title}{An optional title for the map.}

\item{legend}{An optional name for the legend.}

\item{num_colors}{The number of colors to use on the map.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many colors.}

\item{zoom}{An optional vector of countries to zoom in on. Elements of this vector must exactly
match the names of countries as they appear in the "region" column of ?country.regions}

\item{show_states}{Wether to draw state borders.}
}
\description{
The map used is mxmunicipio.map. See ?mxmunicipio.map for
more information.
}
\examples{
df <- df_mxmunicipio
df$value <- df$indigenous
mxmunicipio_choropleth(df)
}
