% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{mxmunicipio_leaflet}
\alias{mxmunicipio_leaflet}
\title{mxmunicipio_leaflet}
\usage{
mxmunicipio_leaflet(df, pal, fillColor, popup, weight = 0.2,
  color = "#555555", opacity = 1, fillOpacity = 0.8, lat = 23.8,
  lng = -102, zoom = 5)
}
\arguments{
\item{zoom}{}
}
\description{
mxmunicipio_leaflet
}

