% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoJSON_data.R
\docType{data}
\name{mxstate.topoJSON}
\alias{mxstate.topoJSON}
\title{Map of the 31 Mexican states plus the Federal District}
\usage{
data(mxstate.topoJSON)
}
\description{
A data.frame which contains a map of all 31 Mexican States plus
the Federal District in topoJSON format.
}
\references{
Downloaded from the "Cartografia Geoestadistica Urbana y
Rural Amanzanada. Planeacion de la Encuesta Intercensal 2015" shapefiles
(https://gist.github.com/diegovalle/aa3eef87c085d6ea034f)
}

