% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{mxstate_choropleth}
\alias{mxstate_choropleth}
\title{Create a state-level choropleth}
\usage{
mxstate_choropleth(df, title = "", legend = "", num_colors = 7,
  zoom = NULL)
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".  Elements in
the "region" column must exactly match how regions are named in the "region" column in ?df_mxstate.}

\item{title}{An optional title for the map.}

\item{legend}{An optional name for the legend.}

\item{num_colors}{The number of colors to use on the map.  A value of 1
will use a continuous scale, and a value in [2, 9] will use that many colors.}

\item{zoom}{An optional vector of countries to zoom in on. Elements of this vector must exactly
match the names of countries as they appear in the "region" column of ?country.regions}
}
\description{
The map used is mxstate.map. See ?mxstate.map for
for more information.
}
\examples{
df <- df_mxstate
df$value <- df$indigenous
mxstate_choropleth(df)
}
