% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{mxstate_leaflet}
\alias{mxstate_leaflet}
\title{mxstate_leaflet}
\usage{
mxstate_leaflet(df, pal, fillColor, popup, weight = 0.2, color = "#555555",
  opacity = 1, fillOpacity = 0.8, lat = 23.8, lng = -102, zoom = 5)
}
\arguments{
\item{zoom}{}
}
\description{
mxstate_leaflet
}

