% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\docType{data}
\name{MXStateChoropleth}
\alias{MXStateChoropleth}
\title{An R6 object for creating state-level choropleths.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MXStateChoropleth
}
\description{
An R6 object for creating state-level choropleths.
}
\keyword{datasets}

